feat: add "stark-debug" feature
