+++ crates/sdk/Cargo.toml
@@ -81,6 +81,8 @@ metrics = [
 ]
 # for guest profiling:
 perf-metrics = ["openvm-circuit/perf-metrics", "openvm-transpiler/function-span"]
+# turns on stark-backend debugger in all proofs
+stark-debug = ["openvm-circuit/stark-debug"]
 test-utils = ["openvm-circuit/test-utils"]
 # performance features:
 # (rayon is always imported because of halo2, so "parallel" feature is redundant)
